+++ src/fobi/models.py
@@ -256,6 +256,9 @@ class FormWizardEntry(models.Model):
 
     user = models.ForeignKey(AUTH_USER_MODEL, verbose_name=_("User"))
     name = models.CharField(_("Name"), max_length=255)
+    title = models.CharField(_("Title"), max_length=255, null=True,
+                             blank=True, help_text=_("Shown in templates if"
+                                                     "available."))
     slug = AutoSlugField(populate_from='name', verbose_name=_("Slug"),
                          unique=True)
     is_public = models.BooleanField(
@@ -275,6 +278,10 @@ class FormWizardEntry(models.Model):
         _("Success page body"), null=True, blank=True,
         help_text=_("Custom message text to display after valid form is "
                     "submitted")
+    )
+    show_all_navigation_buttons = models.BooleanField(
+        _("Show all navigation buttons?"), default=False,
+        help_text=_("Show all navigation buttons.")
     )
     # action = models.CharField(
     #     _("Action"), max_length=255, null=True, blank=True,
@@ -319,20 +326,20 @@ class FormEntry(models.Model):
     :Properties:
 
         - `user` (django.contrib.auth.models.User: User owning the plugin.
-        - `wizard` (str): Form wizard to which the form entry belongs to.
         - `name` (str): Form name.
+        - `title` (str): Form title - used in templates.
         - `slug` (str): Form slug.
         - `description` (str): Form description.
         - `is_public` (bool): If set to True, is visible to public.
         - `is_cloneable` (bool): If set to True, is cloneable.
         - `position` (int): Ordering position in the wizard.
     """
-    # form_wizard_entry = models.ForeignKey(
-    #     FormWizardEntry, verbose_name=_("Form wizard"), null=True, blank=True
-    # )
 
     user = models.ForeignKey(AUTH_USER_MODEL, verbose_name=_("User"))
     name = models.CharField(_("Name"), max_length=255)
+    title = models.CharField(_("Title"), max_length=255, null=True,
+                             blank=True, help_text=_("Shown in templates if"
+                                                     "available."))
     slug = AutoSlugField(
         populate_from='name', verbose_name=_("Slug"), unique=True
     )